+++ hibernate-core/src/test/java/org/hibernate/orm/test/idprops/PropertyNamedIdInEmbeddedIdTest.java
@@ -4,44 +4,69 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.test.idprops;
+package org.hibernate.orm.test.idprops;
 
 import java.io.Serializable;
+import javax.persistence.Embeddable;
+import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.IdClass;
 
 import org.hibernate.testing.TestForIssue;
-import org.hibernate.testing.junit4.BaseCoreFunctionalTestCase;
-import org.junit.Before;
-import org.junit.Test;
+import org.hibernate.testing.orm.junit.DomainModel;
+import org.hibernate.testing.orm.junit.SessionFactory;
+import org.hibernate.testing.orm.junit.SessionFactoryScope;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import static org.hibernate.testing.transaction.TransactionUtil.doInHibernate;
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * @author Gail Badner
  */
-public class PropertyNamedIdInIdClassTest extends BaseCoreFunctionalTestCase {
-	@Override
-	public Class[] getAnnotatedClasses() {
-		return new Class[] { Person.class };
-	}
+@DomainModel(
+		annotatedClasses = PropertyNamedIdInEmbeddedIdTest.Person.class
+)
+@SessionFactory
+public class PropertyNamedIdInEmbeddedIdTest {
 
-	@Before
-	public void setUp() {
-		doInHibernate( this::sessionFactory, session -> {
+	@BeforeEach
+	public void setUp(SessionFactoryScope scope) {
+		scope.inTransaction( session -> {
 			session.persist( new Person( "John Doe", 0 ) );
 			session.persist( new Person( "John Doe", 1 ) );
 			session.persist( new Person( "Jane Doe", 0 ) );
 		} );
 	}
 
+	@AfterEach
+	public void tearDown(SessionFactoryScope scope) {
+		scope.inTransaction(
+				session ->
+						session.createQuery( "delete from Person" ).executeUpdate()
+		);
+	}
+
 	@Test
 	@TestForIssue(jiraKey = "HHH-13084")
-	public void testHql() {
-		doInHibernate( this::sessionFactory, session -> {
-			assertEquals( 2, session.createQuery( "from Person p where p.id = 0", Person.class ).list().size() );
+	public void testHql(SessionFactoryScope scope) {
+		scope.inTransaction( session -> {
+
+			assertEquals(
+					1,
+					session.createQuery( "from Person p where p.id = :id", Person.class )
+							.setParameter( "id", new PersonId( "John Doe", 0 ) )
+							.list()
+							.size()
+			);
+
+			assertEquals(
+					2,
+					session.createQuery( "from Person p where p.id.id = :id", Person.class )
+							.setParameter( "id", 0 )
+							.list()
+							.size()
+			);
 
 			assertEquals( 3L, session.createQuery( "select count( p ) from Person p" ).uniqueResult() );
 
@@ -49,41 +74,25 @@ public class PropertyNamedIdInIdClassTest extends BaseCoreFunctionalTestCase {
 	}
 
 	@Entity(name = "Person")
-	@IdClass(PersonId.class)
 	public static class Person implements Serializable {
-		@Id
-		private String name;
+		@EmbeddedId
+		private PersonId personId;
 
-		@Id
-		private Integer id;
-
-		public Person(String name, Integer id) {
+		public Person(String name, int id) {
 			this();
-			setName( name );
-			setId( id );
-		}
-
-		public String getName() {
-			return name;
-		}
-
-		public Integer getId() {
-			return id;
+			personId = new PersonId( name, id );
 		}
 
 		protected Person() {
 			// this form used by Hibernate
 		}
 
-		protected void setName(String name) {
-			this.name = name;
-		}
-
-		protected void setId(Integer id) {
-			this.id = id;
+		public PersonId getPersonId() {
+			return personId;
 		}
 	}
 
+	@Embeddable
 	public static class PersonId implements Serializable {
 		private String name;
 		private Integer id;
