+++ hibernate-core/src/test/java/org/hibernate/orm/test/idprops/PropertyNamedIdOutOfIdClassTest.java
@@ -4,49 +4,68 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.test.idprops;
+package org.hibernate.orm.test.idprops;
 
 import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.IdClass;
 
-import org.hibernate.testing.junit4.BaseCoreFunctionalTestCase;
-import org.junit.Before;
-import org.junit.Test;
+import org.hibernate.testing.TestForIssue;
+import org.hibernate.testing.orm.junit.DomainModel;
+import org.hibernate.testing.orm.junit.SessionFactory;
+import org.hibernate.testing.orm.junit.SessionFactoryScope;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-import static org.hibernate.testing.transaction.TransactionUtil.doInHibernate;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * @author Gail Badner
  */
-public class PropertyNamedIdOutOfNonJpaCompositeIdTest extends BaseCoreFunctionalTestCase {
-	@Override
-	public Class[] getAnnotatedClasses() {
-		return new Class[] { Person.class };
-	}
+@DomainModel(
+		annotatedClasses = PropertyNamedIdOutOfIdClassTest.Person.class
+)
+@SessionFactory
+public class PropertyNamedIdOutOfIdClassTest {
 
-	@Before
-	public void setUp() {
-		doInHibernate( this::sessionFactory, session -> {
+	@BeforeEach
+	public void setUp(SessionFactoryScope scope) {
+		scope.inTransaction( session -> {
 			session.persist( new Person( "John Doe", 0 ) );
 			session.persist( new Person( "John Doe", 1, 1 ) );
 			session.persist( new Person( "John Doe", 2, 2 ) );
 		} );
 	}
 
-	@Test
-	public void testHql() {
-		doInHibernate( this::sessionFactory, session -> {
-			assertEquals( 1, session.createQuery( "from Person p where p.id = 1", Person.class ).list().size() );
+	@AfterEach
+	public void tearDown(SessionFactoryScope scope) {
+		scope.inTransaction(
+				session ->
+						session.createQuery( "delete from Person" ).executeUpdate()
+		);
+	}
 
+	@Test
+	@TestForIssue(jiraKey = "HHH-13084")
+	public void testHql(SessionFactoryScope scope) {
+		scope.inTransaction( session -> {
+			assertEquals(
+					1,
+					session.createQuery( "from Person p where p.id is null", Person.class ).list().size()
+			);
+			assertEquals(
+					2,
+					session.createQuery( "from Person p where p.id is not null", Person.class ).list().size()
+			);
 			assertEquals( 3L, session.createQuery( "select count( p ) from Person p" ).uniqueResult() );
 		} );
 	}
 
 	@Entity(name = "Person")
+	@IdClass(PersonId.class)
 	public static class Person implements Serializable {
 		@Id
 		private String name;
